pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gateio-websocket-framework"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" }, # 請替換成您的資訊
]
description = "A simple and easy-to-use Python framework for the Gate.io WebSocket API v4."
readme = "README.md"
requires-python = ">=3.7"
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "websocket-client>=1.0.0", # 指定依賴套件與版本
]

[project.urls]
"Homepage" = "https://github.com/your-username/GateioFramework" # 請替換成您的 GitHub Repo URL
"Bug Tracker" = "https://github.com/your-username/GateioFramework/issues" # 請替換